Move FilterSidebar and FilterHeader into separate components, fix layout
